java webstart working, basic support for escape peaks

+++ Peakaboo_CurveFit/src/peakaboo/curvefit/fitting/FittingFunction.java
@@ -16,6 +16,6 @@ public interface FittingFunction {
 	 * @param point the point to evauluate this function at
 	 * @return the result of evaluating this function at the given point
 	 */
-	public double getHeightAtPoint(double point);
+	public float getHeightAtPoint(float point);
 	
 }
